Move save and plot function to utils.external
